requirements.txt
# Core AI Frameworks
langchain==0.2.5
langgraph==0.0.69
langchain-google-genai==1.0.6
langchain-core==0.2.10
langchain-community==0.2.5

# UI and Scheduling
streamlit==1.36.0
APScheduler==3.10.4

# Data Handling
pandas==2.2.2
openpyxl==3.1.3
faker==25.8.0

# Environment and Tooling
python-dotenv==1.0.1
filelock==3.14.0

# Testing
pytest==8.2.2

python-dateutil==2.9.0
